+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,9 +1,9 @@
 import AuthForm from '../../shared/components/AuthForm';
-import { FormWrapper, RedirectLink, Link } from './LoginPage.styled';
+import { FormWrapper } from './LoginPage.styled';
 import FormTitle from '../../shared/components/FormTitile';
 import Section from '../../shared/components/Section/Section';
 import Container from '../../shared/components/Container/Container';
-
+import RedirectLink from '../../shared/components/RedirectLink';
 const LoginPage = () => {
   return (
     <Section>
@@ -11,10 +11,7 @@ const LoginPage = () => {
         <FormWrapper>
           <FormTitle text="Login" />
           <AuthForm isLogin />
-          <RedirectLink>
-            Don't have an account?&nbsp;
-            <Link to="/register">Register</Link>
-          </RedirectLink>
+          <RedirectLink text="Don't have an account?" />
         </FormWrapper>
       </Container>
     </Section>
