+++ src/shared/components/RedirectLink/RedirectLink.jsx
@@ -1,0 +1,16 @@
+import { RedirectText, Link } from './RedirectLink.styled';
+
+const RedirectLink = ({ text, isLogin }) => {
+  return (
+    <RedirectText>
+      {text}&nbsp;
+      {isLogin ? (
+        <Link to="/login">Login</Link>
+      ) : (
+        <Link to="/register">Register</Link>
+      )}
+    </RedirectText>
+  );
+};
+
+export default RedirectLink;
