+++ src/shared/components/RedirectLink/index.js
@@ -1,0 +1,1 @@
+export { default } from './RedirectLink.jsx';
